Add assertions and reorganise files
