+++ src/main/java/emily/app/Launcher.java
@@ -1,0 +1,13 @@
+package emily.app;
+
+import javafx.application.Application;
+
+/**
+ * Launches the program.
+ */
+public class Launcher {
+
+    public static void main(String[] args) {
+        Application.launch(UI.class, args);
+    }
+}
