--- src/test/java/ToDosTest.java
@@ -1,4 +1,0 @@
-package test.java;
-
-public class ToDosTest {
-}
